+++ js/appConfig.js
@@ -8,7 +8,7 @@
 
 const Proj4js = require('proj4');
 const SearchProviders = require('./SearchProviders');
-const CoordinatesUtils = require('../qwc2/MapStore2/web/client/utils/CoordinatesUtils');
+const CoordinatesUtils = require('../qwc2/MapStore2Components/utils/CoordinatesUtils');
 const renderHelp = require('./Help');
 
 Proj4js.defs("EPSG:21781", "+proj=somerc +lat_0=46.95240555555556 +lon_0=7.439583333333333 +k_0=1 +x_0=600000 +y_0=200000 +ellps=bessel +towgs84=674.4,15.1,405.3,0,0,0,0 +units=m +no_defs");