+++ ChangeLog.md
@@ -83,6 +83,7 @@ v0.3 (XXX XX 2017):
 - The print legend functionality requires a assets/templates/legendprint.html file with an element with id="legendcontainer", which will contain the legend graphics. The HTML document title of this template **may** influence the suggested output name when printing to a file (depending on the browser).
 - The enabled MapPlugin components need to be specified in appConfig.js
 - The custom extent in themesConfig.json needs to be specified w.r.t. the crs specified in the theme mapCrs
+- Ported react components to use class definition and ES7 syntax, see i.e. [https://babeljs.io/blog/2015/06/07/react-on-es6-plus](https://babeljs.io/blog/2015/06/07/react-on-es6-plus). **Any custom react components will need to be updated**.
 
 
 v0.2 (Jan 03 2017):